[frontend] files to tsx
