--- opencti-platform/opencti-front/src/private/components/common/stix_domain_objects/StixDomainObjectAttackPatternsKillChain.jsx
@@ -1,373 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import * as R from 'ramda';
-import { graphql } from 'react-relay';
-import Tooltip from '@mui/material/Tooltip';
-import { FileDownloadOutlined, FilterAltOutlined, InvertColorsOffOutlined, ViewColumnOutlined, ViewListOutlined } from '@mui/icons-material';
-import { ProgressWrench } from 'mdi-material-ui';
-import ToggleButtonGroup from '@mui/material/ToggleButtonGroup';
-import ToggleButton from '@mui/material/ToggleButton';
-import IconButton from '@mui/material/IconButton';
-import withStyles from '@mui/styles/withStyles';
-import StixCoreObjectsExports from '../stix_core_objects/StixCoreObjectsExports';
-import SearchInput from '../../../../components/SearchInput';
-import Security from '../../../../utils/Security';
-import { KNOWLEDGE_KNGETEXPORT, KNOWLEDGE_KNUPDATE } from '../../../../utils/hooks/useGranted';
-import StixCoreRelationshipCreationFromEntity from '../stix_core_relationships/StixCoreRelationshipCreationFromEntity';
-import StixDomainObjectAttackPatternsKillChainMatrix from './StixDomainObjectAttackPatternsKillChainMatrix';
-import StixDomainObjectAttackPatternsKillChainLines from './StixDomainObjectAttackPatternsKillChainLines';
-import ExportButtons from '../../../../components/ExportButtons';
-import Filters from '../lists/Filters';
-import FilterIconButton from '../../../../components/FilterIconButton';
-import { export_max_size } from '../../../../utils/utils';
-import inject18n from '../../../../components/i18n';
-
-const styles = (theme) => ({
-  container: {
-    width: '100%',
-    height: '100%',
-    margin: 0,
-    padding: 0,
-  },
-  parameters: {
-    marginBottom: 20,
-    padding: 0,
-  },
-  filtersDialog: {
-    margin: '0 0 20px 0',
-  },
-  icon: {
-    paddingTop: 4,
-    display: 'inline-block',
-    color: theme.palette.primary.main,
-  },
-  text: {
-    display: 'inline-block',
-    flexGrow: 1,
-    marginLeft: 10,
-  },
-  autocomplete: {
-    float: 'left',
-    margin: '5px 10px 0 10px',
-    width: 200,
-  },
-  export: {
-    float: 'right',
-    margin: '0 0 0 20px',
-  },
-  chips: {
-    display: 'flex',
-    flexWrap: 'wrap',
-  },
-  chip: {
-    margin: theme.spacing(1) / 4,
-  },
-});
-
-class StixDomainObjectAttackPatternsKillChain extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      currentModeOnlyActive: false,
-      currentColorsReversed: false,
-      targetEntities: [],
-    };
-  }
-
-  handleToggleModeOnlyActive() {
-    this.setState({ currentModeOnlyActive: !this.state.currentModeOnlyActive });
-  }
-
-  handleToggleColorsReversed() {
-    this.setState({ currentColorsReversed: !this.state.currentColorsReversed });
-  }
-
-  handleAdd(entity) {
-    this.setState({ targetEntities: [entity] });
-  }
-
-  render() {
-    const {
-      t,
-      classes,
-      data,
-      stixDomainObjectId,
-      entityLink,
-      handleSearch,
-      helpers,
-      filters,
-      handleChangeView,
-      searchTerm,
-      currentView,
-      paginationOptions,
-      openExports,
-      handleToggleExports,
-      exportContext,
-      availableFilterKeys,
-      refetch,
-    } = this.props;
-    const { currentColorsReversed, currentModeOnlyActive, targetEntities } = this.state;
-    let csvData = null;
-    if (currentView === 'courses-of-action') {
-      csvData = R.pipe(
-        R.map((n) => n.node.to.coursesOfAction.edges),
-        R.flatten,
-        R.map((n) => n.node),
-      )(data.stixCoreRelationships.edges);
-    }
-    const exportDisabled = targetEntities.length > export_max_size;
-
-    const newExportContext = { ...exportContext, entity_type: 'Attack-Pattern' };
-    const newPaginationOptions = {
-      orderBy: 'name',
-      orderMode: 'desc',
-      filters: {
-        mode: 'and',
-        filters: [
-          {
-            key: 'regardingOf',
-            values: [{
-              key: 'id',
-              values: [stixDomainObjectId],
-            }],
-          },
-        ],
-        filterGroups: [],
-      },
-    };
-
-    return (
-      <>
-        <div
-          className={classes.parameters}
-          style={{ marginTop: -12 }}
-        >
-          <div style={{ float: 'left', marginRight: 20 }}>
-            <SearchInput
-              variant="small"
-              keyword={searchTerm}
-              onSubmit={handleSearch.bind(this)}
-            />
-          </div>
-          <div
-            style={{ float: 'left', display: 'flex', margin: '-6px 4px 0 0' }}
-          >
-            <Tooltip
-              title={
-                currentModeOnlyActive
-                  ? t('Display the whole matrix')
-                  : t('Display only used techniques')
-              }
-            >
-              <span>
-                <IconButton
-                  color={currentModeOnlyActive ? 'secondary' : 'primary'}
-                  onClick={this.handleToggleModeOnlyActive.bind(this)}
-                  size="large"
-                >
-                  <FilterAltOutlined fontSize="medium"/>
-                </IconButton>
-              </span>
-            </Tooltip>
-            <Tooltip
-              title={
-                currentColorsReversed
-                  ? t('Disable invert colors')
-                  : t('Enable invert colors')
-                }
-            >
-              <span>
-                <IconButton
-                  color={currentColorsReversed ? 'secondary' : 'primary'}
-                  onClick={this.handleToggleColorsReversed.bind(this)}
-                  size="large"
-                >
-                  <InvertColorsOffOutlined fontSize="medium"/>
-                </IconButton>
-              </span>
-            </Tooltip>
-            <Filters
-              availableFilterKeys={availableFilterKeys}
-              helpers={helpers}
-              searchContext={{ entityTypes: ['Attack-Pattern'] }}
-            />
-            <FilterIconButton
-              filters={filters}
-              helpers={helpers}
-              styleNumber={2}
-              redirection
-              searchContext={{ entityTypes: ['Attack-Pattern'] }}
-            />
-          </div>
-          <div style={{ float: 'right', margin: 0 }}>
-            <ToggleButtonGroup size="small" color="secondary" exclusive={true}>
-              <Tooltip title={t('Matrix view')}>
-                <ToggleButton onClick={handleChangeView.bind(this, 'matrix')}>
-                  <ViewColumnOutlined
-                    fontSize="small"
-                    color={currentView === 'matrix' ? 'secondary' : 'primary'}
-                  />
-                </ToggleButton>
-              </Tooltip>
-              <Tooltip title={t('Kill chain view')}>
-                <ToggleButton onClick={handleChangeView.bind(this, 'list')}>
-                  <ViewListOutlined
-                    fontSize="small"
-                    color={currentView === 'list' ? 'secondary' : 'primary'}
-                  />
-                </ToggleButton>
-              </Tooltip>
-              <Tooltip title={t('Courses of action view')}>
-                <ToggleButton
-                  onClick={handleChangeView.bind(this, 'courses-of-action')}
-                >
-                  <ProgressWrench
-                    fontSize="small"
-                    color={
-                      currentView === 'courses-of-action'
-                        ? 'secondary'
-                        : 'primary'
-                    }
-                  />
-                </ToggleButton>
-              </Tooltip>
-              {typeof handleToggleExports === 'function' && !exportDisabled && (
-              <Tooltip title={t('Open export panel')}>
-                <ToggleButton
-                  value="export"
-                  aria-label="export"
-                  onClick={handleToggleExports.bind(this)}
-                >
-                  <FileDownloadOutlined
-                    fontSize="small"
-                    color={openExports ? 'secondary' : 'primary'}
-                  />
-                </ToggleButton>
-              </Tooltip>
-              )}
-              {typeof handleToggleExports === 'function' && exportDisabled && (
-              <Tooltip
-                title={`${
-                  t(
-                    'Export is disabled because too many entities are targeted (maximum number of entities is: ',
-                  ) + export_max_size
-                })`}
-              >
-                <span>
-                  <ToggleButton
-                    size="small"
-                    value="export"
-                    aria-label="export"
-                    disabled={true}
-                  >
-                    <FileDownloadOutlined fontSize="small"/>
-                  </ToggleButton>
-                </span>
-              </Tooltip>
-              )}
-            </ToggleButtonGroup>
-            <div className={classes.export}>
-              <ExportButtons
-                domElementId="container"
-                name={t('Attack patterns kill chain')}
-                csvData={csvData}
-                csvFileName={`${t('Attack pattern courses of action')}.csv`}
-              />
-            </div>
-          </div>
-          <div className="clearfix"/>
-        </div>
-        <div className={classes.container}>
-          {currentView === 'list' && (
-          <StixDomainObjectAttackPatternsKillChainLines
-            data={data}
-            entityLink={entityLink}
-            paginationOptions={paginationOptions}
-            onDelete={refetch.bind(this)}
-            searchTerm={searchTerm}
-          />
-          )}
-          {currentView === 'matrix' && (
-          <StixDomainObjectAttackPatternsKillChainMatrix
-            data={data}
-            entityLink={entityLink}
-            searchTerm={searchTerm}
-            handleToggleModeOnlyActive={this.handleToggleModeOnlyActive.bind(
-              this,
-            )}
-            handleToggleColorsReversed={this.handleToggleColorsReversed.bind(
-              this,
-            )}
-            currentColorsReversed={currentColorsReversed}
-            currentModeOnlyActive={currentModeOnlyActive}
-            handleAdd={this.handleAdd.bind(this)}
-          />
-          )}
-          {currentView === 'courses-of-action' && (
-          <StixDomainObjectAttackPatternsKillChainLines
-            data={data}
-            entityLink={entityLink}
-            paginationOptions={paginationOptions}
-            onDelete={refetch.bind(this)}
-            searchTerm={searchTerm}
-            coursesOfAction={true}
-          />
-          )}
-          <Security needs={[KNOWLEDGE_KNUPDATE]}>
-            <StixCoreRelationshipCreationFromEntity
-              entityId={stixDomainObjectId}
-              isRelationReversed={false}
-              paddingRight={220}
-              onCreate={refetch.bind(this)}
-              targetStixDomainObjectTypes={['Attack-Pattern']}
-              paginationOptions={paginationOptions}
-              targetEntities={targetEntities}
-            />
-          </Security>
-          <Security needs={[KNOWLEDGE_KNGETEXPORT]}>
-            <StixCoreObjectsExports
-              open={openExports}
-              exportType='simple'
-              handleToggle={handleToggleExports.bind(this)}
-              paginationOptions={newPaginationOptions}
-              exportContext={newExportContext}
-            />
-          </Security>
-        </div>
-      </>
-    );
-  }
-}
-
-StixDomainObjectAttackPatternsKillChain.propTypes = {
-  data: PropTypes.object,
-  entity: PropTypes.object,
-  entityLink: PropTypes.string,
-  currentView: PropTypes.string,
-  searchTerm: PropTypes.string,
-  handleChangeView: PropTypes.func,
-  handleSearch: PropTypes.func,
-  helpers: PropTypes.object,
-  filters: PropTypes.array,
-  paginationOptions: PropTypes.object,
-  classes: PropTypes.object,
-  t: PropTypes.func,
-  availableFilterKeys: PropTypes.array,
-  refetch: PropTypes.func,
-};
-
-export const stixDomainObjectAttackPatternsKillChainStixCoreRelationshipsQuery = graphql`
-    query StixDomainObjectAttackPatternsKillChainStixCoreRelationshipsQuery(
-        $fromOrToId: [String]
-        $elementWithTargetTypes: [String]
-        $first: Int
-        $filters: FilterGroup
-    ) {
-        ...StixDomainObjectAttackPatternsKillChainContainer_data
-    }
-`;
-
-export default R.compose(
-  inject18n,
-  withStyles(styles),
-)(StixDomainObjectAttackPatternsKillChain);
