+++ opencti-platform/opencti-front/src/private/components/common/stix_domain_objects/StixDomainObjectAttackPatternsKillChainLines.jsx
@@ -170,6 +170,7 @@ class StixDomainObjectAttackPatternsKillChainLines extends Component {
                   <List>
                     {stixCoreRelationship.attackPatterns.map(
                       (attackPattern) => {
+                        console.log('attackPattern', attackPattern);
                         const link = `${entityLink}/relations/${attackPattern.id}`;
                         return (
                           <div key={attackPattern.id}>
